test: move server-as-channel coverage from client to server
