+++ packages/client/test/client.spec.js
@@ -6,10 +6,6 @@ import * as Service from './service.js'
 import { Receipt, Message, CBOR } from '@ucanto/core'
 import { alice, bob, mallory, service as w3 } from './fixtures.js'
 import fetch from '@web-std/fetch'
-// @ts-ignore test-only package is outside client project references
-import * as Server from '@ucanto/server'
-// @ts-ignore test-only package is outside client project references
-import { Schema } from '@ucanto/validator'
 
 test('encode invocation', async () => {
   /** @type {Client.ConnectionView<Service.Service>} */
@@ -142,261 +138,4 @@ test('encode delegated invocation', async () => {
       remove.capabilities
     )
   }
-})
-
-// Create the service instance
-const service = Service.create()
-
-// Define capabilities
-const storeAddCapability = Server.capability({
-  can: 'store/add',
-  with: Server.URI.match({ protocol: 'did:' }),
-  nb: Schema.struct({
-    link: Server.Link.match().optional(),
-  }),
-  derives: (
-    /** @type {{ with: string, nb: { link?: unknown } }} */ claimed,
-    /** @type {{ with: string, nb: { link?: unknown } }} */ delegated
-  ) => {
-    if (claimed.with !== delegated.with) {
-      return Server.fail(
-        `Expected 'with: "${delegated.with}"' instead got '${claimed.with}'`
-      )
-    } else if (
-      delegated.nb.link &&
-      `${delegated.nb.link}` !== `${claimed.nb.link}`
-    ) {
-      return Server.fail(
-        `Link ${
-          claimed.nb.link == null ? '' : `${claimed.nb.link} `
-        }violates imposed ${delegated.nb.link} constraint`
-      )
-    } else {
-      return Server.ok({})
-    }
-  },
-})
-
-const storeRemoveCapability = Server.capability({
-  can: 'store/remove',
-  with: Server.URI.match({ protocol: 'did:' }),
-  nb: Schema.struct({
-    link: Server.Link.match().optional(),
-  }),
-  derives: (
-    /** @type {{ with: string, nb: { link?: unknown } }} */ claimed,
-    /** @type {{ with: string, nb: { link?: unknown } }} */ delegated
-  ) => {
-    if (claimed.with !== delegated.with) {
-      return Server.fail(
-        `Expected 'with: "${delegated.with}"' instead got '${claimed.with}'`
-      )
-    } else if (
-      delegated.nb.link &&
-      `${delegated.nb.link}` !== `${claimed.nb.link}`
-    ) {
-      return Server.fail(
-        `Link ${
-          claimed.nb.link == null ? '' : `${claimed.nb.link} `
-        }violates imposed ${delegated.nb.link} constraint`
-      )
-    } else {
-      return Server.ok({})
-    }
-  },
-})
-
-// Create server with service handlers using Server.provide
-const server = Server.create({
-  id: w3,
-  service: {
-    store: {
-      add: Server.provide(
-        storeAddCapability,
-        async (
-          /** @type {{ capability: unknown, invocation: unknown }} */ {
-            capability,
-            invocation,
-          }
-        ) => {
-          // Call the existing service method with the invocation
-          return await service.store.add(
-            /** @type {Client.Invocation<any>} */ (invocation)
-          )
-        }
-      ),
-      remove: Server.provide(
-        storeRemoveCapability,
-        async (
-          /** @type {{ capability: unknown, invocation: unknown }} */ {
-            capability,
-            invocation,
-          }
-        ) => {
-          // Call the existing service method with the invocation
-          return await service.store.remove(
-            /** @type {Client.Invocation<any>} */ (invocation)
-          )
-        }
-      ),
-    },
-  },
-  codec: CAR.inbound,
-  validateAuthorization: () => ({ ok: {} }),
-})
-
-/**
- * @typedef {{
- *   store: {
- *     add: Client.ServiceMethod<any, any, any>
- *     remove: Client.ServiceMethod<any, any, any>
- *   }
- * }} StoreService
- */
-
-// Use server directly as channel (no HTTP, no mock fetch!)
-/** @type {Client.ConnectionView<StoreService>} */
-const connection = Client.connect({
-  id: w3,
-  channel: server, // 🎯 Server directly as channel - validates delegation chains!
-  codec: CAR.outbound,
-})
-
-test('execute', async () => {
-  const car = await CAR.codec.write({
-    roots: [await CBOR.write({ hello: 'world ' })],
-  })
-
-  const add = Client.invoke({
-    issuer: alice,
-    audience: w3,
-    capability: {
-      can: 'store/add',
-      with: alice.did(),
-      nb: { link: car.cid },
-    },
-    proofs: [],
-  })
-
-  const remove = Client.invoke({
-    issuer: alice,
-    audience: w3,
-    capability: {
-      can: 'store/remove',
-      with: alice.did(),
-      nb: { link: car.cid },
-    },
-  })
-
-  const e1 = await add.execute(connection)
-
-  assert.deepEqual(e1.out, {
-    error: {
-      name: 'UnknownDIDError',
-      message: `DID ${alice.did()} has no account`,
-      did: alice.did(),
-    },
-  })
-
-  // fake register alice
-  service.access.accounts.register(
-    alice.did(),
-    'did:email:alice@web.mail',
-    car.cid
-  )
-
-  const [r1] = await connection.execute(add)
-  assert.deepEqual(r1.out, {
-    ok: {
-      with: alice.did(),
-      link: car.cid,
-      status: 'upload',
-      url: 'http://localhost:9090/',
-    },
-  })
-})
-
-test('execute with delegations', async () => {
-  const car = await CAR.codec.write({
-    roots: [await CBOR.write({ hello: 'world ' })],
-  })
-
-  const add = Client.invoke({
-    issuer: bob,
-    audience: w3,
-    capability: {
-      can: 'store/add',
-      with: bob.did(),
-      nb: { link: car.cid },
-    },
-    proofs: [],
-  })
-
-  const [e1] = await connection.execute(await add.delegate())
-
-  assert.deepEqual(e1.out, {
-    error: {
-      name: 'UnknownDIDError',
-      message: `DID ${bob.did()} has no account`,
-      did: bob.did(),
-    },
-  })
-
-  // fake register alice
-  service.access.accounts.register(bob.did(), 'did:email:bob@web.mail', car.cid)
-
-  const [r1] = await connection.execute(await add.delegate())
-  assert.deepEqual(r1.out, {
-    ok: {
-      with: bob.did(),
-      link: car.cid,
-      status: 'upload',
-      url: 'http://localhost:9090/',
-    },
-  })
-})
-
-test('decode error', async () => {
-  const client = Client.connect({
-    id: w3,
-    channel: server,
-    codec: Codec.outbound({
-      encoders: {
-        'application/car': CAR.request,
-      },
-      decoders: {
-        'application/car+receipt': CAR.response,
-      },
-    }),
-  })
-
-  const car = await CAR.codec.write({
-    roots: [await CBOR.write({ hello: 'world ' })],
-  })
-
-  const add = Client.invoke({
-    issuer: alice,
-    audience: w3,
-    capability: {
-      can: 'store/add',
-      with: alice.did(),
-      nb: { link: car.cid },
-    },
-    proofs: [],
-  })
-
-  const [e1] = await client.execute(await add.delegate())
-
-  assert.deepEqual(
-    {
-      error: {
-        message:
-          "Can not decode response with content-type 'application/vnd.ipld.car' because no matching transport decoder is configured.",
-        // @ts-expect-error
-        name: 'TypeError',
-        error: true,
-      },
-    },
-    e1.out
-  )
 })
